backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Database
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# NLP & ML (optional — platform runs without these in demo mode)
numpy==1.26.4
scikit-learn==1.4.2

# Caching
redis==5.0.4

# File processing
pypdf2==3.0.1
python-docx==1.1.0

# Utils
python-dotenv==1.0.1
loguru==0.7.2